+++ firmware/src/control/navigator.cpp
@@ -53,20 +53,20 @@ Navigator::Navigator(void) {
 /**
  *
  */
-NavOut<float> Navigator::update(const NavIn<float> &in) {
+NavOut<double> Navigator::update(const NavIn<double> &in) {
 
   osalDbgCheck(ready);
 
   auto crosstrack = sphere.crosstrack(in.lat, in.lon);
   auto crs_dist = sphere.course_distance(in.lat, in.lon);
 
-  return NavOut<float>(crosstrack.xtd, crosstrack.atd, crs_dist.dist, crs_dist.crs);
+  return NavOut<double>(crosstrack.xtd, crosstrack.atd, crs_dist.dist, crs_dist.crs);
 }
 
 /**
  *
  */
-void Navigator::loadLine(const NavLine<float> &line) {
+void Navigator::loadLine(const NavLine<double> &line) {
   ready = true;
   sphere.updatePoints(line.latA, line.lonA, line.latB, line.lonB);
 }
